refactor(rooms-api): add modular room controllers and services (availability, CRUD, pricing)
Introduce dedicated controllers and services to replace the monolithic room implementation. This improves separation of concerns and prepares the API for tenant-scoped room features.
