+++ apps/api/src/routers/room.router.ts
@@ -1,13 +1,21 @@
 import { Router } from "express";
-import { roomController } from "../controllers/room.controller.js";
+import { roomCrudController } from "../controllers/room-crud.controller.js";
+import { roomAvailabilityController } from "../controllers/room-availability.controller.js";
+import { roomPricingController } from "../controllers/room-pricing.controller.js";
 import { verifyTokenMiddleware } from "../middlewares/verifyToken.middleware.js";
 import { verifyRoleMiddleware } from "../middlewares/verifyRole.middleware.js";
 import { upload } from "../middlewares/upload.middleware.js";
 
 const router = Router();
 
-router.get("/:roomId/unavailable-dates", roomController.getUnavailableDates);
-router.get("/:roomId/availability", roomController.getRoomAvailability);
+router.get(
+  "/:roomId/unavailable-dates",
+  roomAvailabilityController.getUnavailableDates
+);
+router.get(
+  "/:roomId/availability",
+  roomAvailabilityController.getRoomAvailability
+);
 
 router.use(verifyTokenMiddleware);
 router.use(verifyRoleMiddleware);
@@ -15,26 +23,36 @@ router.use(verifyRoleMiddleware);
 router.post(
   "/property/:propertyId",
   upload.single("imageFile"),
-  roomController.createRoom
+  roomCrudController.createRoom
 );
-router.get("/property/:propertyId", roomController.getRoomsByProperty);
+router.get("/property/:propertyId", roomCrudController.getRoomsByProperty);
 
-router.get("/:roomId", roomController.getRoomById);
-router.put("/:roomId", upload.single("imageFile"), roomController.updateRoom);
-router.delete("/:roomId", roomController.deleteRoom);
+router.get("/:roomId", roomCrudController.getRoomById);
+router.put(
+  "/:roomId",
+  upload.single("imageFile"),
+  roomCrudController.updateRoom
+);
+router.delete("/:roomId", roomCrudController.deleteRoom);
 
-router.post("/:roomId/block", roomController.blockRoomDates);
-router.post("/:roomId/unblock", roomController.unblockRoomDates);
+router.post("/:roomId/block", roomAvailabilityController.blockRoomDates);
+router.post("/:roomId/unblock", roomAvailabilityController.unblockRoomDates);
 
-router.get("/:roomId/price-adjustments", roomController.getPriceAdjustments);
-router.post("/:roomId/price-adjustments", roomController.createPriceAdjustment);
+router.get(
+  "/:roomId/price-adjustments",
+  roomPricingController.getPriceAdjustments
+);
+router.post(
+  "/:roomId/price-adjustments",
+  roomPricingController.createPriceAdjustment
+);
 router.put(
   "/price-adjustments/:adjustmentId",
-  roomController.updatePriceAdjustment
+  roomPricingController.updatePriceAdjustment
 );
 router.delete(
   "/price-adjustments/:adjustmentId",
-  roomController.deletePriceAdjustment
+  roomPricingController.deletePriceAdjustment
 );
 
 export default router;
